Update title and license
Please note that, due to pre-existing ambiguity, .ipynb files in *previous* commits may be used under CC-BY 3.0, although stronger (standard) attribution is expected in this case, with adequate attribution during delivery to enable attendees to access this repository. From this commit onwards, .ipynb files are unambiguously under CC-BY-SA 3.0 and, while more relaxed attribution is required (as stated in README.md), the Share-Alike provisions must be adhered to (as also indicated in the README.md file).